+++ src/utils/test/load-current-session.test.ts
@@ -40,7 +40,12 @@ describe('loadCurrentSession', () => {
 
     const cookieId = '1234-this-is-a-cookie-session-id';
 
-    const session = new Session(cookieId);
+    const session = new Session(
+      cookieId,
+      'test-shop.myshopify.io',
+      'state',
+      true,
+    );
     await expect(
       Context.SESSION_STORAGE.storeSession(session),
     ).resolves.toEqual(true);
@@ -76,7 +81,12 @@ describe('loadCurrentSession', () => {
     } as http.IncomingMessage;
     const res = {} as http.ServerResponse;
 
-    const session = new Session(`test-shop.myshopify.io_${jwtPayload.sub}`);
+    const session = new Session(
+      `test-shop.myshopify.io_${jwtPayload.sub}`,
+      'test-shop.myshopify.io',
+      'state',
+      true,
+    );
     await expect(
       Context.SESSION_STORAGE.storeSession(session),
     ).resolves.toEqual(true);
@@ -140,7 +150,12 @@ describe('loadCurrentSession', () => {
 
     const cookieId = '1234-this-is-a-cookie-session-id';
 
-    const session = new Session(cookieId);
+    const session = new Session(
+      cookieId,
+      'test-shop.myshopify.io',
+      'state',
+      true,
+    );
     await expect(
       Context.SESSION_STORAGE.storeSession(session),
     ).resolves.toEqual(true);
@@ -159,7 +174,12 @@ describe('loadCurrentSession', () => {
 
     const cookieId = ShopifyOAuth.getOfflineSessionId('test-shop.myshopify.io');
 
-    const session = new Session(cookieId);
+    const session = new Session(
+      cookieId,
+      'test-shop.myshopify.io',
+      'state',
+      false,
+    );
     await expect(
       Context.SESSION_STORAGE.storeSession(session),
     ).resolves.toEqual(true);
@@ -185,6 +205,9 @@ describe('loadCurrentSession', () => {
 
     const session = new Session(
       ShopifyOAuth.getOfflineSessionId('test-shop.myshopify.io'),
+      'test-shop.myshopify.io',
+      'state',
+      false,
     );
     await expect(
       Context.SESSION_STORAGE.storeSession(session),